Enhance buf rule to auto download/update binary
We can add some simple logic to check if buf is installed and up to
date, and will no-op the step if everything is already good to go.

We set the buf target as a dependency of the proto rule, and thus any
rules that need to build the binary (such as install or install-global)
will be guaranteed to work.

Note: If the user has their own version of buf installed into a personal
bin folder or the go/bin path, the behaviour will be undefined.